back-end/pyproject.toml
[project]
name = "back-end"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "google-cloud-aiplatform>=1.80.0",
    "python-dotenv>=1.0.1",
    "supabase>=2.13.0",
]
